+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- Allow defining setup and teardown with method names.
 - Allow multiple `setup`, `around` and `teardown` blocks in the same test suite.
 
 ## [0.5.0] - 2025-01-17